# api/openapi.yaml
openapi: 3.0.3
info:
  title: Muasal API
  version: 0.1.0
  description: >-
    The API the web app uses, open to scripts through personal access tokens (FSD §14.3). Send
    `Authorization: Bearer msl_…`; a token acts with its owner's visibility and role, a read-only token only reads
    (and may ask), and each token takes 60 requests a minute, bursting to 120. Browser sessions use the sid cookie
    and must send the app's Origin on writes.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
paths:
...
  /auth/setup:
    post:
      operationId: setupPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetupRequest" }
      responses:
        "204": { description: Password set; every session of the user is revoked. }
        default: { $ref: "#/components/responses/Problem" }
  /auth/setup/account:
    post:
      operationId: getSetupAccount
      tags: [auth]
      description: >-
        The account a live setup link belongs to, so the setup page can name it (MSL-20). The token goes in the body,
        never the URL, to stay out of access logs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: The account.
          content:
            application/json:
              schema:
                type: object
                required: [name, email]
                properties:
                  name: { type: string }
                  email: { type: string }
        default: { $ref: "#/components/responses/Problem" }
  /me:
    get:
      operationId: getMe
      tags: [me]
      responses:
        "200":
          description: The signed-in user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      operationId: updateMe
      tags: [me]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MeUpdate" }
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default: { $ref: "#/components/responses/Problem" }
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      responses:
        "200":
          description: Every user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      operationId: createUser
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserCreate" }
      responses:
        "201":
          description: The user, with a one-time setup link.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedUser" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserUpdate" }
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        default: { $ref: "#/components/responses/Problem" }
  /admin/users/{id}/setup-link:
    post:
      operationId: createSetupLink
      tags: [admin]
      description: Resets the password; the old one stops working and every session ends.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "201":
          description: A new one-time setup link.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SetupLink" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  responses:
    Problem:
      description: An error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Locale:
      type: string
      enum: [id, en]
    User:
      type: object
      required: [id, email, name, is_admin, locale, timezone, disabled, has_password, created_at, last_login_at]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        name: { type: string }
        is_admin: { type: boolean }
        locale: { $ref: "#/components/schemas/Locale" }
        timezone: { type: string, example: Asia/Jakarta }
        disabled: { type: boolean }
        has_password: { type: boolean }
        created_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time, nullable: true }
        notify_prefs: { $ref: "#/components/schemas/NotifyPrefs" }
    UserList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/User" }
...
    SetupRequest:
      type: object
      required: [token, password]
      properties:
        token: { type: string }
        password: { type: string, maxLength: 1024 }
    MeUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        locale: { $ref: "#/components/schemas/Locale" }
        timezone: { type: string }
        current_password: { type: string, maxLength: 1024 }
        new_password: { type: string, maxLength: 1024 }
        notify_prefs: { $ref: "#/components/schemas/NotifyPrefs" }
    UserCreate:
      type: object
      required: [email, name]
      properties:
        email: { type: string, maxLength: 320 }
        name: { type: string, maxLength: 200 }
        is_admin: { type: boolean }
        locale: { $ref: "#/components/schemas/Locale" }
        timezone: { type: string }
    UserUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        is_admin: { type: boolean }
        locale: { $ref: "#/components/schemas/Locale" }
        timezone: { type: string }
        disabled: { type: boolean }
    SetupLink:
      type: object
      required: [url, expires_at]
      properties:
        url: { type: string }
        expires_at: { type: string, format: date-time }
        emailed_to: { type: string, description: With email set up, where the link is being sent (MSL-50). }
    CreatedUser:
      type: object
      required: [user, setup_link]
      properties:
        user: { $ref: "#/components/schemas/User" }
        setup_link: { $ref: "#/components/schemas/SetupLink" }
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
...
    NotifyPrefs:
      type: object
      description: Each event on or off, and browser notifications (§8.10). An absent event is on; browser is off until chosen.
      properties:
        assigned: { type: boolean }
        comment: { type: boolean }
        mention: { type: boolean }
        status: { type: boolean }
        job_done: { type: boolean }
        due: { type: boolean, description: The morning reminder of tickets due today or tomorrow, or just overdue (MSL-52). }
        browser: { type: boolean }
        email: { type: boolean, description: "Also by email, a digest of what is still unread after two minutes, when the server has SMTP set up (MSL-10). Off until chosen." }
...